requirements/requirements.txt
django
psycopg2-binary
dj-database-url
python-decouple
django-webpack4-loader
unipath

easy-thumbnails
django-image-cropping


#django-crispy-forms==1.7.2
#django-floppyforms==1.7.0
#django-extra-views==0.12.0
#django-model-utils==3.1.2
#
#pytz==2020.1
#urllib3
#django-extensions==2.2.9##

#drf-yasg==1.13.0
#djangorestframework==3.9.0
#django-filter==2.0.0
#
#
#django-webpack4-loader
